+++ prep.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 # script name:     prep.sh
-# last modified:   2018/01/07
+# last modified:   2018/06/03
 # sudo:            yes
 
 script_name=$(basename -- "$0")
@@ -24,3 +24,22 @@ apt -y install libnetcdf-dev
 apt -y install python3-pip
 apt -y install python3-venv
 apt -y install libzmq3-dev
+
+# dependencies for python-opencv-headless
+#------------------------------------------------------
+apt -y install libjasper libjasper-dev
+apt -y install libjpeg-dev libtiff5-dev libpng12-dev
+apt -y install libilmbase12
+apt -y install libopenexr22
+apt -y install libgstreamer1.0-0
+apt -y install libavcodec-extra57
+apt -y install libavformat-dev
+apt -y install libilmbase12
+apt -y install libavcodec-dev
+apt -y install libswscale-dev
+apt -y install libv4l-dev
+apt -y install libgtk2.0-dev
+apt -y install libgtk-3-dev
+apt -y install libxvidcore-dev
+apt -y install libx264-dev
+#------------------------------------------------------